+++ src/com/example/calculator/Operator.java
@@ -2,19 +2,19 @@ package com.example.calculator;
 
 public enum Operator { // “미리 정해진 버튼만 있는 리모컨”
     ADD('+') { // 연산을 객체로 만들기: 문자 -> 안전한 타입으로 변환
-        public int apply(int a, int b) {
+        public double apply(double a, double b) {
             return a + b;
         }
     }, SUB('-'){
-        public int apply(int a, int b) {
+        public double apply(double a, double b) {
             return a - b;
         }
     }, MUL('*') {
-        public int apply(int a, int b) {
+        public double apply(double a, double b) {
             return a * b;
         }
     }, DIV('/') {
-        public int apply(int a, int b) {
+        public double apply(double a, double b) {
             if(b == 0) throw new ArithmeticException("0으로 나눌 수 없습니다.");
             return a / b;
         }
@@ -26,7 +26,7 @@ public enum Operator { // “미리 정해진 버튼만 있는 리모컨”
         this.symbol = symbol;
     }
 
-    public abstract int apply (int a, int b); // 연산이 어떻게 계산되는지 정의
+    public abstract double apply (double a, double b); // 연산이 어떻게 계산되는지 정의
 
     public static Operator from (char symbol) { // 사용자 입력을 enum으로 변환
         for (Operator op : values()) {
